+++ frontend/src/components/Like/Like.jsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { useVideo } from "../../contexts";
-import { useDocumentTitle, useToastCleaner } from "../../utils";
+import { useDocumentTitle } from "../../utils";
 import { VideoListing, PlayListHeading } from "../index";
 
 function Like() {
   const {
     state: { likedVideos },
   } = useVideo();
-  useToastCleaner();
   useDocumentTitle("Liked");
+  
   return (
     <div>
       <PlayListHeading
