+++ frontend/src/components/Playlists/PlayLists.jsx
@@ -1,19 +1,15 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import { useVideo } from "../../contexts";
-import {
-  getFilteredList,
-  useDocumentTitle,
-  useToastCleaner,
-} from "../../utils";
-import PlayListView from "./PlayListView";
+import { getFilteredList, useDocumentTitle } from "../../utils";
+import { PlayListView } from "../index";
 import "./PlayLists.css";
 
 function PlayLists() {
   const {
     state: { playlists, likedVideos, savedVideos },
   } = useVideo();
-  useToastCleaner();
+
   useDocumentTitle("Playlists");
   const updatedPlaylists = getFilteredList(playlists);
   return (
