+++ frontend/src/components/Playlists/PlaylistDetail/PlayListDetail.jsx
@@ -6,7 +6,6 @@ import {
   getFilteredList,
   getListDetails,
   useDocumentTitle,
-  useToastCleaner,
 } from "../../../utils";
 import { useVideo } from "../../../contexts";
 import PlayListHeading from "./PlayListHeading";
@@ -17,8 +16,8 @@ function PlayListDetail() {
     dispatch,
   } = useVideo();
   const { playlistId } = useParams();
-  useToastCleaner();
   useDocumentTitle("PlaylistDetail");
+
   const currentPlaylist = getListDetails(playlists, playlistId);
   const filteredVideoList = currentPlaylist
     ? getFilteredList(currentPlaylist?.videoList)
