+++ frontend/src/components/Saved/Saved.jsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { useVideo } from "../../contexts";
-import { useDocumentTitle, useToastCleaner } from "../../utils";
+import { useDocumentTitle } from "../../utils";
 import { VideoListing, PlayListHeading } from "../index";
 
 function Saved() {
   const {
     state: { savedVideos },
   } = useVideo();
-  useToastCleaner();
   useDocumentTitle("Saved");
+
   return (
     <div>
       {savedVideos ? (
