+++ frontend/src/components/index.js
@@ -3,6 +3,8 @@ export { default as VideoListing } from "./Video/VideoListing/VideoListing";
 export { default as VideoDetail } from "./Video/VideoDetail/VideoDetail";
 export { default as PlayListDetail } from "./Playlists/PlaylistDetail/PlayListDetail";
 export { default as PlayListHeading } from "./Playlists/PlaylistDetail/PlayListHeading";
+export { default as PlayListView } from "./Playlists/PlaylistDetail/PlayListView";
+export { default as PlayListModal } from "./Playlists/PlaylistDetail/PlayListModal";
 export { default as PlayLists } from "./Playlists/PlayLists";
 export { default as Like } from "./Like/Like";
 export { default as Saved } from "./Saved/Saved";
@@ -11,3 +13,5 @@ export { default as Toast } from "./Toast/Toast";
 export { default as NotFound } from "./NotFound/NotFound";
 export { handleToast } from "./Toast/Toast";
 export { closeToast } from "./Toast/Toast";
+export { default as ScrollToTop } from "./NotFound/ScrollToTop";
+export { OutlineButton } from "./Video/VideoDetail/VideoActions";
